# apps/api/src/swagger.yml
openapi: 3.0.0
info:
  title: Kitchen OpenAPI
  description: Kitchen StaticSettings description
  version: 1.0.0
  contact:
    name: Vladimir Shabalin
    url: 'https://github.com/mustordont'
    email: mustordont@gmail.com
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ApplicationEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        name:
          type: string
        appPort:
          type: number
        netscalerIP:
          type: string
        netscalerURL:
          type: string
        description:
          type: string
      required:
        - id
        - name
        - appPort
        - netscalerIP
        - netscalerURL
        - description
    HostEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        applications:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationEntity'
        ip:
          type: string
        dnsname:
          type: string
        os:
          type: string
        cpu:
          type: number
        ram:
          type: number
        hdd:
          type: number
        auth:
          type: string
        citrix:
          type: string
        description:
          type: string
      required:
        - id
        - applications
        - ip
        - dnsname
        - os
        - cpu
        - ram
        - hdd
        - auth
        - citrix
        - description
    NetworkEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        production:
          type: boolean
        base:
          type: string
        mask:
          type: string
        vos:
          type: number
        vlan:
          type: number
        description:
          type: string
        hosts:
          type: array
          items:
            $ref: '#/components/schemas/HostEntity'
      required:
        - id
        - production
        - base
        - mask
        - vos
        - vlan
        - description
        - hosts
    GroupEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        name:
          type: string
        applications:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationEntity'
      required:
        - id
        - name
        - applications
    KuberEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        namespace:
          type: string
        applications:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationEntity'
      required:
        - id
        - namespace
        - applications
    DomainEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        name:
          type: string
        networks:
          type: array
          items:
            $ref: '#/components/schemas/NetworkEntity'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupEntity'
        kubers:
          type: array
          items:
            $ref: '#/components/schemas/KuberEntity'
      required:
        - id
        - name
        - networks
        - groups
        - kubers
...
    CertificateJobEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        emails:
          type: string
        status:
          type: object
      required:
        - id
        - emails
        - status
    CertEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        validFrom:
          type: string
        validTo:
          type: string
        subject:
          type: string
        alertDate:
          type: string
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/CertificateJobEntity'
      required:
        - id
        - validFrom
        - validTo
        - subject
        - alertDate
        - jobs
...
    CertFileEntity:
      type: object
      properties:
        id:
          type: number
        created:
          type: string
        updated:
          type: string
        name:
          type: string
        file:
          type: string
        status:
          type: object
        certs:
          type: array
          items:
            $ref: '#/components/schemas/CertEntity'
      required:
        - id
        - name
        - file
        - status
        - certs
paths:
...
  /api/landscape/domain:
    get:
      operationId: DomainController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DomainEntity'
      tags: &ref_1
        - landscape
    post:
      operationId: DomainController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainEntity'
      tags: *ref_1
  '/api/landscape/domain/{id}/network':
    post:
      operationId: DomainController_addNetwork
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainEntity'
      tags: *ref_1
  '/api/landscape/domain/{id}/group':
    post:
      operationId: DomainController_addGroup
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainEntity'
      tags: *ref_1
  '/api/landscape/domain/{id}/kuber':
    post:
      operationId: DomainController_addKuber
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KuberEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainEntity'
      tags: *ref_1
  '/api/landscape/domain/{id}':
    get:
      operationId: DomainController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainEntity'
      tags: *ref_1
    put:
      operationId: DomainController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainEntity'
      tags: *ref_1
    delete:
      operationId: DomainController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_1
...
  /api/landscape/host/search:
    get:
      operationId: HostController_search
      parameters:
        - name: dnsname
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HostEntity'
      tags: &ref_5
        - landscape
...
  '/api/landscape/host/{id}':
    get:
      operationId: HostController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostEntity'
      tags: *ref_5
    put:
      operationId: HostController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostEntity'
      tags: *ref_5
    delete:
      operationId: HostController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_5
...
  '/api/landscape/application/{id}':
    get:
      operationId: ApplicationController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationEntity'
      tags: *ref_6
    put:
      operationId: ApplicationController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationEntity'
      tags: *ref_6
    delete:
      operationId: ApplicationController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_6
...
  '/api/certificate/file/{id}/certificate':
    get:
      operationId: CertificateFileController_addCertificate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertFileEntity'
      tags: *ref_9
...
  /api/certificate:
    get:
      operationId: CertificateController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CertEntity'
      tags: &ref_11
        - certificates
    post:
      operationId: CertificateController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertEntity'
      tags: *ref_11
  '/api/certificate/{id}':
    get:
      operationId: CertificateController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertEntity'
      tags: *ref_11
    put:
      operationId: CertificateController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertEntity'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertEntity'
      tags: *ref_11
    delete:
      operationId: CertificateController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_11